+++ packages/fictoan-react/src/components/Form/RadioButton/RadioGroup.tsx
@@ -30,6 +30,8 @@ export const RadioGroup = React.forwardRef(
             equalizeWidth,
             required,
             disabled,
+            size,
+            columns,
             ...props
         }: RadioGroupProps,
         ref: React.Ref<HTMLDivElement>
@@ -50,6 +52,10 @@ export const RadioGroup = React.forwardRef(
             classNames.push(`equalise-width`);
         }
 
+        if (columns) {
+            classNames.push(`with-columns`);
+        }
+
         return (
             <FormItem
                 label={label}
@@ -57,6 +63,7 @@ export const RadioGroup = React.forwardRef(
                 helpText={helpText}
                 errorText={errorText}
                 required={required}
+                size={size}
             >
                 <Element
                     as="div"
@@ -65,6 +72,7 @@ export const RadioGroup = React.forwardRef(
                     classNames={classNames}
                     role="radiogroup"
                     aria-label={label}
+                    style={columns ? { gridTemplateColumns: `repeat(${columns}, 1fr)` } : undefined}
                     {...props}
                 >
                     {options.map((option, index) => {